[project @ standardize password retrieval code] both get_checked_password and get_password now take a JID and password if the password (which defaults to None) is not None just return it (this makes it easier to make functions that allow you to pass in a password but still will prompt for it if not provided)
if the password is none both will construct a prompt from the jabber id
and use getpass to ask for the password.

I also renamed all the instances of passwd to password because its
bad to spell things differently in two different places.

Finally I removed the copy of toJID from benderjab.py as there
was no reason to have two copies of it.