[project @ make base BenderJab Bot class function as a daemon] *preliminary* this patch will break all of my examples because I've radically changed how the bot is configured. The factory is gone, and the bot itself knows how to read the config file.
the bot now also has a main function which handles parsing
the command line arguments, and calling start/stop/restart.

Additionally it takes advantage of the createDaemon function I found
to detach from the controlling terminal.

Unfortunately all of this code is bascially unix specific. 

I'll try to update the examples to match. I'm going to accept 
the unix specificity for the moment.