Add in support for sending to email addresses.
It will use the jabber ID as the email sender address.

Also create a logger and attach it to the bot class instead
of just always using the root object.

The logging still needs improvement as it doesn't use the python logging
configuration file, but given the design of one config file for multiple
JIDs I couldn't quickly think of how to modify it.