Use decorators to configure arguments and commands for bots.
This is an experimental branch as I try to see if its plausable
to use a customized version of argparse to create the bot
command parser.

Since the parser was intended to return strings to create messages
I'm not sure how it should work if we want to send directed Iq
or Presence messages.

Also I'm not sure how a command should handle callbacks
for actions it generated.